Integer serialization fix for 32 bit platform compatibility (#52)
Motivation:

[RFC 9651](https://www.ietf.org/rfc/rfc9651.html) permits serializing an
integer in the range of -999,999,999,999,999 to 999,999,999,999,999
inclusive ([Section
4.1.4-2.1.1](https://www.ietf.org/rfc/rfc9651.html#section-4.1.4-2.1.1)).
As `RFC9651BareItem` takes values `integer` or `date` with an associated
value of type `Int`, that range is implicitly more constrained on 32-bit
platforms to a subset range of `Int32.min` to `Int32.max` inclusive.

Modifications:

- Adjust `RFC9651BareItem` to take values `integer` or `date` with an
associated value of type `Int64` instead, which is sufficient for the
permitted range.

Result:

Compatibility for the Integer and Date Structured Types on 32-bit
platforms.

+++ Sources/RawStructuredFieldValues/FieldParser.swift
@@ -232,7 +232,7 @@ extension StructuredFieldValueParser {
     }
 
     private mutating func _parseAnIntegerOrDecimal(isDate: Bool) throws -> RFC9651BareItem {
-        var sign = 1
+        var sign = Int64(1)
         var type = IntegerOrDecimal.integer
 
         if let first = self.underlyingData.first, first == asciiDash {
@@ -304,7 +304,7 @@ extension StructuredFieldValueParser {
         case .integer:
             // This intermediate string is sad, we should rewrite this manually to avoid it.
             // This force-unwrap is safe, as we have validated that all characters are ascii digits.
-            let baseInt = Int(String(decoding: integerBytes, as: UTF8.self), radix: 10)!
+            let baseInt = Int64(String(decoding: integerBytes, as: UTF8.self), radix: 10)!
             let resultingInt = baseInt * sign
 
             if isDate {
